feat: componentizacao da wallet screen
