+++ lib/app_widget.dart
@@ -1,9 +1,9 @@
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
-import 'package:projeto_criptos/src/presenter/screens/wallet/wallet_page.dart';
+import 'package:projeto_criptos/src/presenter/screens/wallet/wallet_screen.dart';
 
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+class AppWidget extends StatelessWidget {
+  const AppWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
