+++ lib/src/presenter/screens/wallet/widgets/asset_list_tile.dart
@@ -1,0 +1,79 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:projeto_criptos/src/presenter/provider/visible_provider.dart';
+
+import '../../../../../utils/currency_formater.dart';
+import '../../../../domain/entities/asset.dart';
+
+class AssetListTile extends StatefulHookConsumerWidget {
+  const AssetListTile({
+    Key? key,
+    required this.asset,
+  }) : super(key: key);
+
+  final AssetEntity asset;
+
+  @override
+  ConsumerState<AssetListTile> createState() => _AssetListTileState();
+}
+
+class _AssetListTileState extends ConsumerState<AssetListTile> {
+  @override
+  Widget build(BuildContext context) {
+    var visible = ref.watch(visibleProvider.state);
+    return ListTile(
+      leading: CircleAvatar(
+        radius: 20,
+        backgroundColor: Colors.transparent,
+        backgroundImage: Image.asset(widget.asset.icon).image,
+      ),
+      title: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(child: Text(widget.asset.symbol)),
+          visible.state
+              ? Text(
+                  realFormatter.format(widget.asset.coinBalance * widget.asset.price),
+                  style: const TextStyle(
+                    fontWeight: FontWeight.w400,
+                    fontSize: 18,
+                  ),
+                )
+              : Container(
+                  width: 110,
+                  height: 18,
+                  decoration: BoxDecoration(
+                    color: Colors.grey.shade300,
+                    borderRadius: BorderRadius.circular(5),
+                  ),
+                ),
+          const Padding(padding: EdgeInsets.only(left: 10)),
+          const Icon(Icons.keyboard_arrow_right),
+        ],
+      ),
+      subtitle: Row(
+        children: [
+          Text(
+            widget.asset.name,
+            style: TextStyle(
+              fontWeight: FontWeight.w500,
+              color: Colors.grey.shade600,
+            ),
+          ),
+          const Spacer(),
+          visible.state
+              ? Text("${widget.asset.coinBalance.toString()} ${widget.asset.symbol}")
+              : Container(
+                  width: 60,
+                  height: 15,
+                  decoration: BoxDecoration(
+                    color: Colors.grey.shade300,
+                    borderRadius: BorderRadius.circular(5),
+                  ),
+                ),
+          const SizedBox(width: 35),
+        ],
+      ),
+    );
+  }
+}
